feat!: tremor v2 \n\n BREAKING CHANGE: incompatible w/ prev versions (#339)
This major release is the first step towards a production-ready version of Tremor. Over the past few months, we have rewritten the library to make it fit for the future. We added the long-awaited exposure of className and support for other HTML attributes, enabling you to overwrite our root styles or extend them with Tailwind CSS.

What's new:
◆ Override and extend root styles with tailwind classes
◆ All HTML attributes extended
◆ Ref forwarding
◆ Bundle size decreased by 25%

tremor.so/changelog